v2.25.0 — Planification des tournées : 3 modes opérationnels, contraintes 6 h, garde de saturation, photos CAV chauffeur (#132)
Moteur de temps pur partagé (budget 6 h, pause déjeuner au centre hors travail,
retours de vidage vehicle-aware, 9 réglages AdminPredictive), estimation stockée
+ refus 409 sur les modes modèle/manuel/association, garde de saturation
(priorité dure + saturation-risks + alerte dashboard réparée), CRUD complet des
modèles de tournées + page /route-templates, wizard réparé (CavPicker,
EstimationPanel), propositions hebdo multi-véhicules, photos des CAV côté
chauffeur (fraîcheur 6 mois paramétrable, photo-public avec périmètre véhicule,
mobile 3 états, offline non bloquant).

Correctifs prouvés sur PostgreSQL 16 réel : collect-public 500 (42P08),
started_at absent du parcours mobile, effets de clôture absents de la route
mobile (completion-effects.js partagé), compte de service « chauffeur », KPI
absentéisme (ROUND double precision), garde de compilation des 183 sources.

Jest 112 suites / 2233 tests verts, Vitest mobile 53/53, builds web + mobile OK,
base neuve re-prouvée (213 tables, idempotence x3).

Au déploiement : deploy.sh update (migration idempotente + seed « chauffeur »),
aucun paramétrage requis.

+++ backend/src/routes/employees.js
@@ -1158,7 +1158,7 @@ router.get('/kpi/absenteisme', authorize('ADMIN', 'RH', 'MANAGER'), async (req, 
              COALESCE(SUM(wh.hours_worked) FILTER (WHERE wh.type IN ('normal','training','absence','sick')), 0)::numeric(10,2) AS heures_totales,
              ROUND(
                100.0 * SUM(wh.hours_worked) FILTER (WHERE wh.type IN ('absence','sick'))::numeric
-               / NULLIF(SUM(wh.hours_worked) FILTER (WHERE wh.type IN ('normal','training','absence','sick')), 0),
+               / NULLIF(SUM(wh.hours_worked) FILTER (WHERE wh.type IN ('normal','training','absence','sick'))::numeric, 0),
                2
              ) AS taux_pct
       FROM employees e
